Rename the command line parameter '-time' -> '-seconds'

+++ sbk-api/src/main/java/io/sbk/api/impl/SbkPerformance.java
@@ -296,7 +296,7 @@ final public class SbkPerformance implements Performance {
                 }
             }
             this.bytes += bytes;
-            this.totalLatency += latency * events;
+            this.totalLatency += (long) latency * events;
             this.maxLatency = Math.max(this.maxLatency, latency);
         }
     }
